wip mcts

+++ DUGENET_YOUSFI_ZHENG_PROGJVAVANCEE_4A/Assets/Scripts/NodeMCTS.cs
@@ -6,8 +6,8 @@ public class NodeMCTS
 {
     public mapSimulation currrentGameState;
 
-    public int nbWin = 0;
-    public int nbMove = 0;
+    public float nbWin = 0;
+    public float nbMove = 0;
 
     
     public NodeMCTS parent;
